Recreate docker-compose setup in kubernetes
my-app changes
    Renamed to app
    Image uses php:7.4-fpm
    Port changed to 9000 (fpm port)

Added app-loadbalancer
    Uses port 80

The public folder is directly copied into the loadbalancer image, while
the entire source folder is copied into the app image. This way the
content that nginx needs to be able to know when to forward to PHP is
available while PHP has all the content it needs to process the request.