Use nginx template rather than static conf file
The kubernetes pod should use the loopback address in order to
connect nginx to the php-fpm process, while docker-compose should use
the name of the container started. Due to this, the nginx conf will have
to be different, but luckily for us nginx official image can handle
templates which replaces env variables with their value and moves them
into the conf directory.